Update braille_alpha-numeric_characters.scad
added extra numerical value to the braille_characters list, to allow the upper and lower case characters to use the same code without being converted from one to another.  added logic to check the first value of the list and then the second to match the character.  added function to create the dots as spherical cap based on a height and width.