Update braille_alpha-numeric_characters.scad
removed function flatten_list_list, uncomment code from function generate_line_list.  add code to check if a list is empty and if not generate a character map for it.  uncomment code to flatten character maps.  add code to make the back plate conform to the length of individual lines (hull) and round the corners (minkowski).  removed code for verified_line_list